# Optimize boolean dictionary