--- a/dictionary_amd64.go
+++ b/dictionary_amd64.go
@@ -30,13 +30,6 @@ func dictionaryLookup32bits(dict []uint32, indexes []int32, rows array, size, of
 //go:noescape
 func dictionaryLookup64bits(dict []uint64, indexes []int32, rows array, size, offset uintptr) errno
 
-func (d *booleanDictionary) lookup(indexes []int32, rows array, size, offset uintptr) {
-	checkLookupIndexBounds(indexes, rows)
-	for i, j := range indexes {
-		*(*bool)(rows.index(i, size, offset)) = d.index(j)
-	}
-}
-
 func (d *int32Dictionary) lookup(indexes []int32, rows array, size, offset uintptr) {
 	checkLookupIndexBounds(indexes, rows)
 	dict := *(*[]uint32)(unsafe.Pointer(&d.values))