@achille-roussel 2022-06-07
Since this is a small change, I'll merge into the parent branch to avoid risking conflicts, this can then be reviewed as part of #214